requirements.txt
langchain
langchain_community
openai
chromadb
tiktoken
pypdf2
fastapi
uvicorn
pytest
python-dotenv~=1.1.0
langchain_openai
pypdf
sentence-transformers
langchain_community
python-multipart
langchain_chroma
langchain_huggingface
langchain_ollama
pydantic
openai-whisper
requests
yt-dlp
langchain_core